Fix function-like macro definition
Function like macros must have no space between the macro name and the opening parenthesis that surrounds the arguments list.
Thankfully, this macro isn't used anywhere.

This requirement in C and C++ is probably one reason codestyle or coding style guidelines or conventions for function definitions and invocations should not include a space between the function name and opening parenthesis of the arguments list but RefindPlus does usually include a space...