Add alternative log macros
For debugging the release version, without including new files, create simple macros that will log messages to the screen. It's not pretty, but can help find where a hang occurs.
The macros require that the format string be a constant string literal which is generally a good idea (and actually normal printf style functions have a warning -Wformat-nonliteral for that but we can't use it for edk2 functions because edk2 uses different % characters than stdlib)