make egDecodeAny non-static
so it can be used by external functions to decode image data. In a future commit of load_icns.c, it will be used to load PNG icons from an .icns file.
Also log when an image type is decoded successfully so we can know what type of data an image file contains.