fix LOG2 call in BdsLibConnectAllDriversToAllControllersEx
I made this mistake during the last merge of upstream code. LOG2 is a macro and requires that the prefix and suffix for MsgLog are string literals - not expressions - so they can be concatenated to the MsgLog string.
Anyway, according to previous commits, the line feed should be in the suffix instead of the prefix in order to satisfy the logging guideline that says log messages should end with a line feed.