simplify InitScroll
The calling function (menu style) should know how many items are visible (MaxVisible) and pass that to InitScroll.

The calculations for MaxVisible that were in InitScroll are moved to the menu styles where they belong so you can make different menu styles in the future without needing to add another calculation to InitScroll.
- TextMenuStyle (for all menus when AllowGraphicsMode == FALSE) This function already had a correct MaxVisible calculation.
- GraphicsMenuStyle (for sub menus when AllowGraphicsMode == TRUE) This function is rewritten in a future commit to correct several calculations including MaxVisible.
- MainMenuStyle (for main menu when AllowGraphicsMode == TRUE) The calculation for MaxVisible is taken from InitScroll.