Simplify back_selected images
Only the alpha should be anti-aliased. The edge pixels of the image should not be antialiased so that the final anti-aliased color has contributions only from the destination pixels (where the image will be drawn) and the edge pixels of the image.

If the edge pixels of the image were anti-aliased, then the surrounding color that the edge pixels were anti-aliased with (usually white) would contribute to the final pixels which is not good if the destination pixels were of a color that is unlike the surrounding color of the image.

In the case of these back_selected images which are a single gray color, the image data can just be a filled rectangle of that gray. The Alpha will contain the shape of the image and have the anti-aliasing info.

A filled rectangle has greater compression so the file size of the *.png and the egemb_*_.h files is smaller.

A note about the images (I don't know if these were intended):
1) Both are gray but are slightly different grays.
2) The small image has an opaque alpha (no background color will show through) while the big image has a partially transparent image (some of the background color will show through).

Preview.app can be used to test the images - it has an option to change the background color that they are drawn on.