use a new flag to decide if image is to be freed
Instead of trying to use a duplication of the logic that causes CompImage to be created to determine if the CompImage needs to be freed, just set a flag when CompImage is created. Much simpler. If the flag doesn't get set then we know that CompImage doesn't need to be freed.