fix unhide tags with same name
Fixed a bug in the Managing Hidden Tags UI where unhiding a legacy option with the same name of a firmware option causes both to be unhidden.
Example from Parallels Desktop UEFI:

Legacy options:
- Boot Legacy (BIOS) OS from Primary Master Harddisk
- Boot Legacy (BIOS) OS from Primary Slave CDROM

Firmware options:
- 0000 Reboot to Primary Master Harddisk
- 0001 Reboot to Primary Slave CDROM

In the above example, if both "Primary Master Harddisk" items existed in the hidden tags lists, then unhiding one of them would actually unhide both of them.
The problem is that the unhide code checks all four lists of hidden tags (Tags, Tools, Legacy, Firmware) for a match to the selected item instead of just the list that the selected item came from.
The simple fix is to store the number of the list that the item came from in the menu entry (use the Row field), then only the list indicated by the Row field number will be affected.

Additional fixes:
- In RemoveInvalidFilenames, VarName is unused. Remove it.