fix missing "Unable to Comply" messages for tools
Added "Manage Firmware Boot Order", "Show Bootscreen", and "Clean NVRAM" to the list of built-in in tools so they get the "Unable to Comply" message when trying to hide them.

Added a new column "Builtin Tool" to the tags list to indicate all the tools that should get this message.
The TAGS_BUILTIN #define is used to put that list of tools in a switch statement.

Having all the tags info in a single file like this would have probably helped catch this issue. Some other things shown by this arrangement:
- There are multiple tools that use the same shortcut key.
- Tool Name shown in the log does not always match Menu Title shown in the UI.