add option to include external hidden icons
Currently, RefindPlus does not attempt to load hidden icons from external volumes. There should be an option to change this behaviour.

A new option, "all_hidden_icons", is added to the config. "all" means that no exclusion is applied when this is set. This flag is separate from ignore_hidden_icons and prefer_hidden_icons which are used in the process of deciding which icon is used in a menu entry and not in the process of loading volume icons.

A message in the log appears if reading an icon for a volume is skipped. It has been changed to indicate what option can be changed to change the behaviour.
"Skipped 'VolName' ... Not Internal Volume ... Config Setting is Not Active:- 'all_hidden_icons'"

A similar message is changed so that the name of the option is output instead of the name of the variable:
"Skipped Loading .VolumeIcons ... Config Setting is Active:- 'ignore_hidden_icons'"